refactor: Dedicated views in dedicated files
