+++ SwissTransferFeatures/SettingsView/Cells/SettingsCell.swift
@@ -21,6 +21,7 @@ import STResources
 import SwiftUI
 import SwissTransferCore
 
+/// A generic Setting cell that works well with enums from STCore
 struct SettingsCell<Content: View>: View {
     let title: String
     let subtitle: String
@@ -52,107 +53,6 @@ struct SettingsCell<Content: View>: View {
                     Image(asset: rightIconAsset)
                         .iconSize(.large)
                 }
-            }
-        }
-    }
-}
-
-/// A view that tracks the notifications settings
-struct NotificationsSettingsCell<Content: View>: View {
-    @AppStorage(UserDefaults.shared.key(.notificationsNewTransfers))
-    private var newTransfers = DefaultPreferences.notificationsNewTransfers
-
-    @AppStorage(UserDefaults.shared.key(.notificationsDownloadInProgress))
-    private var downloadInProgress = DefaultPreferences.notificationsDownloadInProgress
-
-    @AppStorage(UserDefaults.shared.key(.notificationsFinishedTransfers))
-    private var finishedTransfers = DefaultPreferences.notificationsFinishedTransfers
-
-    @AppStorage(UserDefaults.shared.key(.notificationsDownloadTransfers))
-    private var downloadTransfers = DefaultPreferences.notificationsDownloadTransfers
-
-    @AppStorage(UserDefaults.shared.key(.notificationsFailedTransfers))
-    private var failedTransfers = DefaultPreferences.notificationsFailedTransfers
-
-    @AppStorage(UserDefaults.shared.key(.notificationsExpiredTransfers))
-    private var expiredTransfers = DefaultPreferences.notificationsExpiredTransfers
-
-    @State private var subtitle: String = NotificationSettings().enabledNotificationLabel
-
-    private let leftIconAsset = STResourcesAsset.Images.bell
-
-    private let title: String = STResourcesStrings.Localizable.settingsOptionNotifications
-
-    @ViewBuilder var destination: () -> Content
-
-    var body: some View {
-        NavigationLink(destination: destination) {
-            HStack(spacing: IKPadding.small) {
-                Image(asset: leftIconAsset)
-                    .iconSize(.large)
-
-                VStack(alignment: .leading) {
-                    Text(title)
-                        .lineLimit(1)
-                        .foregroundStyle(Color.ST.textPrimary)
-                        .font(.ST.headline)
-                    Text(subtitle)
-                        .lineLimit(1)
-                        .foregroundStyle(Color.ST.textSecondary)
-                        .font(.ST.callout)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .onChange(of: [newTransfers,
-                               downloadInProgress,
-                               finishedTransfers,
-                               downloadTransfers,
-                               failedTransfers,
-                               expiredTransfers]) { newValue in
-                    let allEnabled = newValue.allSatisfy { $0 }
-                    if allEnabled {
-                        self.subtitle = STResourcesStrings.Localizable.settingsAllNotifications
-                    } else {
-                        self.subtitle = STResourcesStrings.Localizable.settingsCustomizedNotifications
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct SingleLabelSettingsCell: View {
-    let title: String
-    var rightIconAsset: STResourcesImages?
-
-    var body: some View {
-        HStack(spacing: IKPadding.small) {
-            Text(title)
-                .lineLimit(1)
-                .foregroundStyle(Color.ST.textPrimary)
-                .font(.ST.headline)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            if let rightIconAsset {
-                Image(asset: rightIconAsset)
-                    .iconSize(.medium)
-            }
-        }
-    }
-}
-
-struct AboutSettingsCell: View {
-    let title: String
-    let subtitle: String
-
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                Text(title)
-                    .foregroundStyle(Color.ST.textPrimary)
-                    .font(.ST.headline)
-                Text(subtitle)
-                    .foregroundStyle(Color.ST.textSecondary)
-                    .font(.ST.callout)
             }
         }
     }
